Fix uint32_t/int "mismatch"
The std::atomic_init(<uint32_t>, 1) fails on some systems with a warning
that "int and uint32_t won't fit through the std::atomic_init()
template".

Added some explicit conversion from "1" to "uint32_t".

+++ src/block.cc
@@ -34,7 +34,7 @@ bool block::init_dummy( void ) {
     static bool did_init{ false };
     if( did_init )
         return did_init;
-    std::atomic_init(&dummy_counter, 1);
+    std::atomic_init(&dummy_counter, (uint32_t)1);
     did_init = true;
     return did_init;
 }
